refactor: moved navbar and category components, to the modules folder
