+++ src/modules/home/ui/components/footer.tsx
